add existing check for all terminal script

+++ install/terminal/app-yazi.sh
@@ -1,13 +1,17 @@
-if [ "$OMAKUB_OS_ID" == "arch" ]; then
-  paru -S --noconfirm --needed yazi
-elif [ "$OMAKUB_OS_ID" == "ubuntu" ]; then
-  git clone https://github.com/sxyazi/yazi.git ~/repos/yazi
+if command -v yazi &>/dev/null; then
+  echo "Yazi is already installed."
+else
+  if [ "$OMAKUB_OS_ID" == "arch" ]; then
+    paru -S --noconfirm --needed yazi
+  elif [ "$OMAKUB_OS_ID" == "ubuntu" ]; then
+    git clone https://github.com/sxyazi/yazi.git ~/repos/yazi
 
-  cd ~/repos/yazi || return 1
-  cargo build --release --locked
-  sudo mv target/release/yazi target/release/ya /usr/local/bin/
-  cd - || return 1
+    cd ~/repos/yazi || return 1
+    cargo build --release --locked
+    sudo mv target/release/yazi target/release/ya /usr/local/bin/
+    cd - || return 1
 
-  # Remove yazi repo
-  rm -rf ~/repos/yazi
+    # Remove yazi repo
+    rm -rf ~/repos/yazi
+  fi
 fi
